+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/dialog/DefaultCalendarDialogBuilder.java
@@ -80,7 +80,7 @@ final class DefaultCalendarDialogBuilder extends AbstractDialogBuilder<CalendarD
 						.owner(owner)
 						.locationRelativeTo(locationRelativeTo)
 						.title(title)
-						.onShown(dialog -> calendarPanel.requestCurrentDayButtonFocus())
+						.onShown(dialog -> calendarPanel.requestInputFocus())
 						.onOk(() -> okPressed.set(true))
 						.show();