+++ changelog.md
@@ -7,6 +7,7 @@ Codion Change Log
 ### is.codion.swing.common.ui
 - CalendarPanel.Builder.locale() added.
 - CalendarPanel now takes the first day of week according to locale into account.
+- CalendarPanel now uses labels instead of buttons to represent days.
 
 ## 0.18.10
 ### is.codion.common.core